pyproject.toml
[tool.poetry]
name = "etl-duckdb"
version = "0.1.0"
description = "Implementando ETL com duckdb e Python"
authors = ["Nicolini <nicolini.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyarrow = "^17.0.0"
pandas = "^2.2.2"
polars = "^1.5.0"
duckdb = "^1.0.0"
magic-duckdb = "^0.1.27"
ipykernel = "^6.29.5"
jupyter = "^1.0.0"
plotly-express = "^0.4.1"
gdown = "^5.2.0"
streamlit = "^1.37.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.32"
pydrive = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
